Added clickable Card hand for UI and reworked how player data is handled

+++ Assets/Scripts/Configurations/UnitConfiguration.cs
@@ -9,13 +9,13 @@ namespace TakeArms.Configurations
     [CreateAssetMenu(fileName = "new UnitConfig", menuName = "Configurations/Unit/Default", order = 0)]
     public class UnitConfiguration : SerializedScriptableObject, IRepositoryItem
     {
-        public ulong id;
+        public int id;
         [ShowInInspector]
         public UnitData data = new UnitData();
 
         public GameObject prefab;
 
-        public ulong ID => id;
+        public int ID => id;
 
         public virtual void InitObject(Transform parentObject)
         {
